% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_if_bad_combs.R
\name{bad_range_warning}
\alias{bad_range_warning}
\title{Internal: Function that constructs a warning message for the user when
  \code{\link{range_checker}} is TRUE.}
\usage{
bad_range_warning(outlier_combs)
}
\description{
Internal: Function that constructs a warning message for the user when
  \code{\link{range_checker}} is TRUE.
}
\keyword{internal}
