% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{disc_trans}
\alias{disc_trans}
\title{Internal: Transform discrete predictions into a usable df}
\usage{
disc_trans(predictions, family, type, model)
}
\description{
Internal: Transform discrete predictions into a usable df
}
