% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{distreg_checker}
\alias{distreg_checker}
\title{Check if model class is supported}
\usage{
distreg_checker(x)
}
\arguments{
\item{x}{Model object or model object in quoted form, e.g. "mymodel"}
}
\description{
This function is a quick way to find out whether a specific model class is supported.
}
\details{
At the moment, the following classes are supported:
\itemize{
\item gamlss
\item bamlss
\item betareg
}
}
