% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_function.R
\name{ex_f}
\alias{ex_f}
\title{External Function Implementer}
\usage{
ex_f(pred_params, unquotedfun)
}
\description{
This function exists to extend \code{plot_moments} such that an external
function, which is user-written, can be included. Thus, the user can see the
impact of a variable on a self-defined measure, like the Gini Index.
}
\keyword{internal}
