% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_specific.R
\name{expl_vars}
\alias{expl_vars}
\title{Internal: Function to obtain all explanatory variables used to fit
  a model, without the dependent variables}
\usage{
expl_vars(model)
}
\description{
Internal: Function to obtain all explanatory variables used to fit
  a model, without the dependent variables
}
\keyword{internal}
