% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fac_check}
\alias{fac_check}
\title{Factor Checker}
\usage{
fac_check(DF)
}
\description{
Checks whether some factor was unwantedly converted to an ordered factor
which rhandsontable sometimes does
}
\keyword{internal}
