% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fac_equ}
\alias{fac_equ}
\title{Factor Equalizer}
\usage{
fac_equ(base_df, pred_df)
}
\description{
Function that takes the levels of a df's factors and puts them to a second df
  (used for predictions). Returns a data.frame
}
\keyword{internal}
