% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fam_fun_getter}
\alias{fam_fun_getter}
\title{Obtain d&p&q functions}
\usage{
fam_fun_getter(fam_name, type)
}
\description{
Takes a family name and what kind of function you want and gives the right one back
}
\keyword{internal}
