% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{fam_obtainer}
\alias{fam_obtainer}
\title{Internal: Family obtainer}
\usage{
fam_obtainer(model)
}
\description{
Gets the right family (in characters) from a given model
}
\examples{
# Generating data
data_fam <- model_fam_data(fam_name = "BE")
# Fit model
library("gamlss")
beta_model <- gamlss(BE ~ norm2 + binomial1,
  data = data_fam, family = BE())
distreg.vis:::fam_obtainer(model = beta_model)
}
\keyword{internal}
