% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_specific.R
\name{gamlss_data_cleaner}
\alias{gamlss_data_cleaner}
\title{GAMLSS expl_data cleaner}
\usage{
gamlss_data_cleaner(temp_df)
}
\description{
This checks whether we have spline column names and/or duplicate columns
}
\keyword{internal}
