% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{has.moments}
\alias{has.moments}
\title{Internal: Distributional Moments implementation checker}
\usage{
has.moments(fam_name)
}
\description{
Is the moment function of a given distribution family implemented? Meaning:
will \code{plot_moments()} work?
}
\keyword{internal}
