% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.2d}
\alias{is.2d}
\title{Internal: Function to check whether the modeled response is bivariate}
\usage{
is.2d(name)
}
\description{
Internal: Function to check whether the modeled response is bivariate
}
\keyword{internal}
