% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.continuous.R
\name{is.continuous}
\alias{is.continuous}
\title{Internal: Function to check whether a family is continuous or not}
\usage{
is.continuous(family)
}
\description{
Internal: Function to check whether a family is continuous or not
}
