% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.discrete}
\alias{is.discrete}
\title{Internal: Discrete Distribution Checker}
\usage{
is.discrete(name)
}
\description{
Check whether a given distribution is fully discrete.
}
\keyword{internal}
