% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.distreg.fam}
\alias{is.distreg.fam}
\title{Internal: Is distreg family}
\usage{
is.distreg.fam(name)
}
\description{
Check whether a given distribution is a distributional regression family
}
\details{
See which classes are currently supported at \link{distreg_checker}.
}
\keyword{internal}
