% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.implemented}
\alias{is.implemented}
\title{Internal: Distribution Implementation Checker}
\usage{
is.implemented(name)
}
\description{
Is the distribution implemented in the distreg.vis framework?
}
\keyword{internal}
