% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{limits}
\alias{limits}
\title{Internal: Get plot limits for a predicted distribution.}
\usage{
limits(predictions, family, times_sd = 3)
}
\description{
Returns a data.frame
}
\keyword{internal}
