% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{lims_getter}
\alias{lims_getter}
\title{Internal: Upper- and lower limit of distribution getter}
\usage{
lims_getter(fam_name)
}
\description{
Obtain the theoretical upper and lower limits of the distribution. Only
necessary if the distribution has limits
}
\keyword{internal}
