% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fam_data.R
\name{model_fam_data}
\alias{model_fam_data}
\title{Create a dataset to fit models with all possible families in bamlss}
\usage{
model_fam_data(nrow = 500, seed = 1408)
}
\description{
Create a dataset to fit models with all possible families in bamlss
}
