% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Return First two Moments of a distributions's parameters}
\usage{
moments(par, fam_name)
}
\arguments{
\item{par}{Parameters of the modeled distribution in a data.frame form.
Can be Output of \code{\link{preds}}, for example.}

\item{family}{One of \code{\link{family.bamlss}} family objects' "family"
description in character form. E.g. \code{"gaussian"} or \code{"beta"}.}
}
\description{
Return First two Moments of a distributions's parameters
}
