% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_trans.R
\name{mult_trans}
\alias{mult_trans}
\title{Internal: Function to transform multinomial predictions}
\usage{
mult_trans(predictions, model)
}
\description{
This function exists solely to transform predictions of the multinomial dist.
Transforms odds into probabilities to get into each class.
}
\keyword{internal}
