% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{palette_getter}
\alias{palette_getter}
\title{Internal: Get colour palettes for 3D plots}
\usage{
palette_getter(name = "default")
}
\description{
Internal: Get colour palettes for 3D plots
}
\keyword{internal}
