% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{pdf_cdf_getter}
\alias{pdf_cdf_getter}
\title{Internal: PDF and CDF getter}
\usage{
pdf_cdf_getter(model)
}
\description{
Obtain the right PDF and CDF for the modeled distribution for a given model class and distribution
}
\keyword{internal}
