% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{pdfcdf_continuous}
\alias{pdfcdf_continuous}
\title{Internal: Create the pdf/cdf for continuous covariates}
\usage{
pdfcdf_continuous(lims, funs, type, p_m, palette, depvar)
}
\description{
Returns a plot
}
\keyword{internal}
