% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moments.R
\name{plot_multinom_exp}
\alias{plot_multinom_exp}
\title{Internal: Plot function as sub-case to plot_moments especially for
  multinomial family}
\usage{
plot_multinom_exp(model, int_var, pred_data, m_data)
}
\description{
Internal: Plot function as sub-case to plot_moments especially for
  multinomial family
}
\keyword{internal}
