% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds.R
\name{preds}
\alias{preds}
\title{Handle predictions in bamlss case (including intercept as variable)}
\usage{
preds(model, newdata)
}
\description{
Handle predictions in bamlss case (including intercept as variable)
}
