% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds.R
\name{preds_transformer}
\alias{preds_transformer}
\title{Sample transformer}
\usage{
preds_transformer(samp_list, newdata)
}
\description{
This functions transforms the bamlss samples from a list for every parameter
to a list for every prediction. This makes it easier to use the moments()
function.
}
\keyword{internal}
