% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_limits.R
\name{quants}
\alias{quants}
\title{Get quantile limits of a distribution}
\usage{
quants(fam_name, pred_params)
}
\description{
Get the quantile limits of a distribution, depending on the predicted parameters.
}
\keyword{internal}
