% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_if_bad_combs.R
\name{range_checker}
\alias{range_checker}
\title{Internal: Function that checks whether chosen covariate combinations
  are in the range of original data. Returns true when there is a cov
  comb outside of data.}
\usage{
range_checker(orig_data, newdata)
}
\description{
Internal: Function that checks whether chosen covariate combinations
  are in the range of original data. Returns true when there is a cov
  comb outside of data.
}
\keyword{internal}
