% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_pmvnorm.R
\name{real_pmvnorm}
\alias{real_pmvnorm}
\title{Internal: Real cumulative density function of multivariate normal dist}
\usage{
real_pmvnorm(y, par)
}
\description{
Internal: Real cumulative density function of multivariate normal dist
}
\keyword{internal}
