% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moments.R
\name{reshape_into_long}
\alias{reshape_into_long}
\title{Internal: Reshape into Long Format}
\usage{
reshape_into_long(
  preds_intvar,
  pred_data,
  int_var,
  int_params,
  samples,
  coltype
)
}
\arguments{
\item{preds_intvar}{A data.frame with the moments as columns and the splitted
\code{int_var} as rows (default 100 values from min to max). There are
three ways which this data.frame can look like. First, as the mean of the
moments. secondly, as the upper and thirdly as the lower quantiles of the
moments.}
}
\description{
Internal: Reshape into Long Format
}
\keyword{internal}
