% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{type_getter}
\alias{type_getter}
\title{Internal: Limit type getter}
\usage{
type_getter(fam_name)
}
\description{
Get the limit type depending on \code{distreg.vis::dists}.
}
\keyword{internal}
