% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_related_helper.R
\name{disc_trans}
\alias{disc_trans}
\title{Internal: Transform discrete predictions into a usable df}
\usage{
disc_trans(pred_params, fam_name, type, model, lims)
}
\description{
Internal: Transform discrete predictions into a usable df
}
\keyword{internal}
