% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{distreg_checker}
\alias{distreg_checker}
\title{Check if object is a distributional regression object}
\usage{
distreg_checker(x)
}
\description{
Check if object is a distributional regression object
}
\details{
At the moment, the following classes are supported:
\itemize{
\item gamlss
\item bamlss
\item betareg
}
}
\keyword{internal}
