% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dists}
\alias{dists}
\title{Information about supported and not yet supported distribution families}
\format{An object of class \code{data.frame} with 124 rows and 8 columns.}
\usage{
dists
}
\description{
A dataset containing all of bamlss' exported and gamlss.dist families. This
is the backbone of the package; whether you can use a family or not depends
on this dataset.
}
\keyword{datasets}
