% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.stuff.R
\name{is.2d}
\alias{is.2d}
\title{Internal: Function to check whether the modeled response is bivariate}
\usage{
is.2d(family, links)
}
\description{
Internal: Function to check whether the modeled response is bivariate
}
\keyword{internal}
