% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.continuous}
\alias{is.continuous}
\title{Internal: Continuous/Mixed Distribution checker}
\usage{
is.continuous(name)
}
\description{
Check whether a given distribution is at least partly continuous (could be
mixed as well).
}
\keyword{internal}
