% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_df_related.R
\name{is.implemented}
\alias{is.implemented}
\title{Internal: Distribution Implementation Checker}
\usage{
is.implemented(fam_name)
}
\description{
Is the distribution generally implemented in the distreg.vis framework?
Meaning: Does \code{plot_dist()} work?
}
\keyword{internal}
