% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_specific.R
\name{model_data}
\alias{model_data}
\title{Model data getter}
\usage{
model_data(model, dep = FALSE, varname = NULL)
}
\arguments{
\item{model}{A gamlss or bamlss object}

\item{dep}{If TRUE, then only the dependent variable is returned}
}
\description{
Get the data from the models
}
\keyword{internal}
