% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{pdfcdf_discrete}
\alias{pdfcdf_discrete}
\title{Internal: Create the pdf/cdf for discrete covariates}
\usage{
pdfcdf_discrete(pred_params, palette, fam_name, type, model, lims, depvar)
}
\description{
Returns a plot
}
\keyword{internal}
