% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Function to plot the predicted distributions (with ggplot2) with
the intercept as another column of the dataframe}
\usage{
plot_dist(model, predictions, palette = "default", type = "pdf")
}
\description{
Function to plot the predicted distributions (with ggplot2) with
the intercept as another column of the dataframe
}
