% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Function to plot the predicted distributions (with ggplot2)}
\usage{
plot_dist(model, newdata, intercept = TRUE)
}
\description{
Function to plot the predicted distributions (with ggplot2)
}
