% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moments.R
\name{plot_moments}
\alias{plot_moments}
\title{Plot function: Display the influence of a covariate}
\usage{
plot_moments(model, int_var, pred_data)
}
\arguments{
\item{model}{A fitted model on which the plots are based.}

\item{int_var}{The variable for which influences of the moments shall be
graphically displayed. Has to be in character form.}

\item{pred_data}{Combinations of covariate data, sometimes also known as "newdata",
including the variable of interest, which will be ignored in later
processing.}
}
\description{
This function takes a dataframe of predictions with one row per
  prediction and one column for every explanatory variable.
  Then, those predictions are held constant while one specific variable
  is varied over it's whole range (min-max). Then, the constant
  variables with the varied interest variables are predicted and plotted
  against the expected value and the variance of the underlying
  distribution
}
